% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisarControleSemestral.R
\name{analisarControleSemestral}
\alias{analisarControleSemestral}
\title{Analisa os dados de controle semestral de acordo com o VMP de cada parâmetro}
\usage{
analisarControleSemestral(periodo, regiao = NULL, uf = NULL, municipios = NULL)
}
\arguments{
\item{periodo}{numeric vector Selecao do ou dos anos de referencia, a partir de 2014.}

\item{regiao}{string vector Selecao da ou das regioes geograficas.}

\item{uf}{string vector Selecao da ou das unidades federativas.}

\item{municipios}{string vector Selecao do ou dos municipios.}
}
\value{
Data frame do conjunto de dados de controle semestral com o total de analises
realizadas e o percentual de amostras acima do VMP para cada parâmetro
}
\description{
Cria uma tabela dos valores de referencia com os VMPs da nova norma de potabilidade
e da norma antiga. Assim, para valores ate 2021 (nao incluso), o VMP considerado e da antiga norma.
De 2021 em diante sao considerados os novos valores de VMP, bem como a retirada
ou inclusao de parametros (deve ser ajustado quando o Sisagua for atualizado
com o modulo de controle semestral, bem como as unidades para etilbenzeno, xilenos e tolueno).
}
\details{
Essa funcao carrega os dados de controle semestral e os analisa de acordo com
a norma de potabilidade vigente no periodo selecionado
}
\note{
O periodo de dados e um campo obrigatorio e inicia em 2014 ate o ano atual.
Os parametros regiao, unidade_federativa e municipios devem sempre ser inseridos com letra
maiuscula e entre aspas, conforme exemplo.
}
\examples{
\dontrun{
df5 <- analisarControleSemestral(periodo = 2023, municipios = "AMPARO")
}

}
