% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ciano.R
\name{download_ciano}
\alias{download_ciano}
\title{Download dos dados de Vigilância - Cianotoxinas e Cianobactérias}
\usage{
download_ciano(periodo, regiao = NULL, unidade_federativa = NULL)
}
\arguments{
\item{periodo}{Selecao do ou dos anos de referencia, a partir de 2014.}

\item{regiao}{Selecao da ou das regioes geograficas.}

\item{unidade_federativa}{Selecao da ou das unidades federativas.}
}
\value{
Data frame do conjunto de dados com filtros a partir dos parametros da funcao
}
\description{
Download dos dados de Vigilância - Cianotoxinas e Cianobactérias
}
\details{
Essa funcao da acesso aos dados do monitoramento da qualidade da agua
para consumo humano realizado pelo setor saude, contemplando os resultados
das analises de cianobacterias e cianotoxinas.
}
\note{
O periodo de dados e um campo obrigatorio e inicia em 2014 ate o ano atual.
Os parametros regiao e unidade_federativa devem sempre ser inseridos com letra
maiuscula e entre aspas, conforme exemplo.
}
\examples{
\dontrun{
df5 <- download_ciano(regiao = "NORTE")
df6 <- download_ciano(unidade_federativa = "CE")
}

}
