% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cmdp.R
\name{download_cmdp}
\alias{download_cmdp}
\title{Function download_cmdp (Controle mensal demais parametros)}
\usage{
download_cmdp(periodo, regiao = NULL, unidade_federativa = NULL)
}
\arguments{
\item{periodo}{Selecao do ou dos anos de referencia, a partir de 2014.}

\item{regiao}{Selecao da ou das regioes geograficas.}

\item{unidade_federativa}{Selecao da ou das unidades federativas.}
}
\value{
Data frame do conjunto de dados com filtros a partir dos parametros da funcao
}
\description{
Function download_cmdp (Controle mensal demais parametros)
}
\note{
O periodo de dados e um campo obrigatorio e inicia em 2014 ate o ano atual.
Os parametros regiao e unidade_federativa devem sempre ser inseridos com letra
maiuscula e entre aspas, conforme exemplo.
}
\examples{
df <- download_cmdp(c(2019:2020), "NORTE", c("AM", "PA", "RO"))
\dontrun{
df5 <- download_cmdp(regiao = "NORTE")
df6 <- download_cmdp(unidade_federativa = "CE")
}

}
